+++ README.md
@@ -1,1 +1,9 @@
 # Certified
+
+start backend
+
+```bash
+cd backend
+
+fastapi dev main.py
+```
