+++ backend/routers/questions.py
@@ -1,7 +1,8 @@
 from typing import List
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
-from ..data import database, schemas, crud
+from ..data import database, schemas
+from ..data.respository import exam_repository as repository
 
 
 router = APIRouter(prefix="/questions", tags=["questions"])
@@ -9,7 +10,7 @@ router = APIRouter(prefix="/questions", tags=["questions"])
 
 @router.get("/{exam_id}", response_model=List[schemas.Question])
 def get_all_questions(exam_id: int, db: Session = Depends(database.get_db)):
-    return crud.get_questions(db, exam_id)
+    return repository.get_questions(db, exam_id)
 
 
 @router.post("/{exam_id}", response_model=schemas.Question)
@@ -18,4 +19,4 @@ def add_questions_with_file(
     question: schemas.QuestionCreate,
     db: Session = Depends(database.get_db),
 ):
-    return crud.create_question(db, exam_id, question)
+    return repository.create_question(db, exam_id, question)
